renamed routing folder and fixed footer styles to pull them from MyStyles.js
